Strip modifier keys of modifiers.
For some reason modifiers are applied to modifier keys being held down when a mode is loaded.
I suspect this is a problem with xcb, as that's where we get key state.
This strips modifiers of modifier keys. I don't see this breaking anything since modifiers
with modifier keys isn't normally applied and is nonesense from the beginning.

This fixes #5288.