Check if subscribe event type is supported in AnyEvent::I3
Add simple `if exists' construct in the subscribe function. This
prevents a somewhat cryptic warnings such as these:

Use of uninitialized value $type in hash element at /usr/share/perl5/AnyEvent/I3.pm line 309.

We still warn the user, but it is much clearer as to what the cause is.

Signed-off-by: Wesley Schwengle <wesleys@opperschaap.net>